+++ src/realm/list.hpp
@@ -144,8 +144,15 @@ public:
     template <typename Func>
     void find_all(value_type value, Func&& func) const
     {
-        if (update())
+        if (update()) {
+            if constexpr (std::is_same_v<T, Mixed>) {
+                if (value.is_null()) {
+                    // if value is null then we find also all the unresolved links with a O(n lg n) scan
+                    find_all_mixed_unresolved_links(std::forward<Func>(func));
+                }
+            }
             m_tree->find_all(value, std::forward<Func>(func));
+        }
     }
 
     inline const BPlusTree<T>& get_tree() const
@@ -258,6 +265,17 @@ protected:
         m_tree->create();
         REALM_ASSERT(m_tree->is_attached());
         return true;
+    }
+
+    template <class Func>
+    void find_all_mixed_unresolved_links(Func&& func) const
+    {
+        for (size_t i = 0; i < m_tree->size(); ++i) {
+            auto mixed = m_tree->get(i);
+            if (mixed.is_unresolved_link()) {
+                func(i);
+            }
+        }
     }
 };
 
@@ -602,7 +620,6 @@ inline void Lst<T>::do_clear()
     m_tree->clear();
 }
 
-
 template <typename U>
 inline Lst<U> Obj::get_list(ColKey col_key) const
 {
@@ -655,7 +672,14 @@ inline T Lst<T>::get(size_t ndx) const
     if (ndx >= current_size) {
         throw std::out_of_range("Index out of range");
     }
-    return m_tree->get(ndx);
+
+    auto value = m_tree->get(ndx);
+    if constexpr (std::is_same_v<T, Mixed>) {
+        // return a null for mixed unresolved link
+        if (value.is_type(type_TypedLink) && value.is_unresolved_link())
+            return Mixed{};
+    }
+    return value;
 }
 
 template <class T>
@@ -663,6 +687,18 @@ inline size_t Lst<T>::find_first(const T& value) const
 {
     if (!update())
         return not_found;
+
+    if constexpr (std::is_same_v<T, Mixed>) {
+        if (value.is_null()) {
+            auto ndx = m_tree->find_first(value);
+            auto size = ndx == not_found ? m_tree->size() : ndx;
+            for (size_t i = 0; i < size; ++i) {
+                if (m_tree->get(i).is_unresolved_link())
+                    return i;
+            }
+            return ndx;
+        }
+    }
     return m_tree->find_first(value);
 }